+++ CHANGELOG.rst
@@ -214,6 +214,8 @@ Added
   `#868 <https://github.com/openego/eGon-data/issues/868>`_
 * Add distribution of CHP plants for eGon100RE
   `#851 <https://github.com/openego/eGon-data/issues/851>`_
+* Add sanity checks for residential electricity loads
+  `#915 <https://github.com/openego/eGon-data/pull/915>`_
 
 .. _PR #159: https://github.com/openego/eGon-data/pull/159
 .. _PR #703: https://github.com/openego/eGon-data/pull/703